requirements.txt
annotated-types==0.6.0
anyio==4.3.0
exceptiongroup==1.2.1
fastapi==0.110.2
greenlet==3.0.3
idna==3.7
pydantic==2.7.0
pydantic_core==2.18.1
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.37.2
typing_extensions==4.10.0
requests==2.32.3
psycopg2-binary==2.9.7